Se agrego el indexer y se movieron los models incluyendo los de test a db/models
