+++ src/entities/accounts/account.services.js
@@ -1,12 +1,12 @@
-const Account = require("./account.model");
+const Account = require("../../db/models/account.model");
 const serviceBuilder = require("../../../libs/serviceBuilder");
 
 const { getAll, getOne, create, update, remove } = serviceBuilder(Account);
 
 module.exports = {
-  getAllAccounts: getAll,
-  getAccount: getOne,
-  createAccount: create,
-  updateAccount: update,
-  removeAccount: remove
+    getAllAccounts: getAll,
+    getAccount: getOne,
+    createAccount: create,
+    updateAccount: update,
+    removeAccount: remove
 };
