+++ src/entities/offers/offer.services.js
@@ -1,13 +1,13 @@
-const Offer = require("./offers.model");
+const Offer = require("../../db/models/offers.model");
 
 const serviceBuilder = require("../../libs/serviceBuilder");
 
 const { getAll, getOne, create, update, remove } = serviceBuilder(Offer);
 
 module.exports = {
-  getAllOffers: getAll,
-  getOffer: getOne,
-  createOffer: create,
-  updateOffer: update,
-  removeOffer: remove
+    getAllOffers: getAll,
+    getOffer: getOne,
+    createOffer: create,
+    updateOffer: update,
+    removeOffer: remove
 };
