+++ src/entities/orders/order.services.js
@@ -1,13 +1,13 @@
-const Order = require("./order.model");
+const Order = require("../../db/models/order.model");
 
 const serviceBuilder = require("../../../libs/serviceBuilder");
 
 const { getAll, getOne, create, update, remove } = serviceBuilder(Order);
 
 module.exports = {
-  getAllOrders: getAll,
-  getOrder: getOne,
-  createOrder: create,
-  updateOrder: update,
-  removeOrder: remove
+    getAllOrders: getAll,
+    getOrder: getOne,
+    createOrder: create,
+    updateOrder: update,
+    removeOrder: remove
 };
