+++ test/mocks/entities/order.mock.js
@@ -1,18 +1,18 @@
 var faker = require("faker");
 
-const Order = require("../../../src/entities/orders/order.model");
+const Order = require("../../../src/db/models/order.model");
 
 const defaultArgumets = () => ({
-  location: {
-    type: "Point",
-    coordinates: [faker.random.number(), faker.random.number()],
-  },
-  address: faker.address.direction(),
-  products: [faker.commerce.productName(), faker.commerce.productName()],
-  services: [faker.commerce.productName(), faker.commerce.productName()],
-  deliverDetails: faker.lorem.paragraph(),
+    location: {
+        type: "Point",
+        coordinates: [faker.random.number(), faker.random.number()],
+    },
+    address: faker.address.direction(),
+    products: [faker.commerce.productName(), faker.commerce.productName()],
+    services: [faker.commerce.productName(), faker.commerce.productName()],
+    deliverDetails: faker.lorem.paragraph(),
 });
 
 module.exports = async function userMocker(args) {
-  return await Order.create({ ...defaultArgumets(), ...args });
+    return await Order.create({...defaultArgumets(), ...args });
 };
