+++ test/mocks/entities/state.mock.js
@@ -1,11 +1,11 @@
 var faker = require("faker");
 
-const State = require("../../../src/entities/states/state.model");
+const State = require("../../../src/db/models/state.model");
 
 const defaultArgumets = () => ({
-  description: faker.lorem.paragraph(),
+    description: faker.lorem.paragraph(),
 });
 
 module.exports = async function userMocker(args) {
-  return await State.create({ ...defaultArgumets(), ...args });
+    return await State.create({...defaultArgumets(), ...args });
 };
